Separate live send execution state contracts
