+++ src/PlateauResoniteLink/Targets/Resonite/LiveSendRunContext.cs
@@ -1,0 +1,8 @@
+namespace PlateauResoniteLink.Targets.Resonite;
+
+internal sealed record LiveSendRunContext(
+    LiveSendRunPlan Plan,
+    CreatedSlot DatasetRootSlot,
+    CreatedSlot DatasetAssetsRootSlot,
+    CreatedSlot CommonAssetsRootSlot,
+    CompositeCityObjectBaker? CityObjectBaker);
