+++ src/PlateauResoniteLink/Targets/Resonite/TerrainTextureAssetCache.cs
@@ -1,0 +1,13 @@
+using System;
+
+namespace PlateauResoniteLink.Targets.Resonite;
+
+internal sealed class TerrainTextureAssetCache
+{
+    public AsyncInFlightResultCache<string, SharedTerrainTextureAsset> AssetsByMeshCode { get; } = new();
+}
+
+internal sealed record SharedTerrainTextureAsset(
+    Uri TextureUri,
+    CreatedComponent TextureComponent,
+    CreatedComponent MainTexturePropertyBlockComponent);
